[docs] 로그인 필요한 API에 Basic header 추가 (#29)
