--- backend/src/main/java/dev/tripdraw/config/SwaggerConfig.java
@@ -1,24 +1,0 @@
-package dev.tripdraw.config;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.servers.Server;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        final Info info = new Info()
-                .title("트립드로우 서버 API 명세서")
-                .version("v0.0.1")
-                .description("API Description");
-        return new OpenAPI()
-                .addServersItem(new Server().url("/"))
-                .components(new Components())
-                .info(info);
-    }
-}
